pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "batch-llm"
version = "2.0.0"
description = "A flexible framework for processing multiple LLM requests efficiently with support for PydanticAI agents, direct API calls, and custom factories"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Geoff Davis", email = "geoff@keksi.ai"}
]
keywords = [
    "llm",
    "batch-processing",
    "pydantic-ai",
    "async",
    "parallel",
    "ai",
    "machine-learning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
pydantic-ai = [
    "pydantic-ai>=0.0.1",
]
gemini = [
    "google-genai>=0.1.0",
]
all = [
    "pydantic-ai>=0.0.1",
    "google-genai>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pydantic-ai>=0.0.1",  # Required for tests
]

[project.urls]
Homepage = "https://github.com/yourusername/batch-llm"
Documentation = "https://github.com/yourusername/batch-llm#readme"
Repository = "https://github.com/yourusername/batch-llm"
Issues = "https://github.com/yourusername/batch-llm/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/batch_llm"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["batch_llm"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselected by default, run with -m slow)",
]
addopts = "-m 'not slow'"  # Skip slow tests by default
